extracted erp_service_double from erp_test
Also create just one service object per erptest instance
